Convert negotiation and trickle tests to vitest
