--- test/common.js
@@ -1,3 +1,0 @@
-// Re-export from TypeScript-generated dist
-export * from '../dist/test/test/common.js'
-export { default } from '../dist/test/test/common.js'
