+++ test/multistream.js
@@ -1,15 +1,14 @@
 import common from './common.js'
 import Peer from '../index.js'
-import test from 'tape'
+import { test, expect } from 'vitest'
 
 test('multistream', function (t) {
-  if (!process.browser) return t.end()
+  if (!process.browser) return
   if (common.isBrowser('ios')) {
-    t.pass('Skip on iOS emulator which does not support this reliably') // iOS emulator issue #486
+    //  // iOS emulator issue #486
     t.end()
     return
   }
-  t.plan(20)
 
   const peer1 = new Peer({
     initiator: true,
@@ -25,31 +24,30 @@ test('multistream', function (t) {
   const receivedIds = {}
 
   peer1.on('stream', function (stream) {
-    t.pass('peer1 got stream')
+    // 
     if (receivedIds[stream.id]) {
-      t.fail('received one unique stream per event')
+      throw new Error('')
     } else {
       receivedIds[stream.id] = true
     }
   })
   peer2.on('stream', function (stream) {
-    t.pass('peer2 got stream')
+    // 
     if (receivedIds[stream.id]) {
-      t.fail('received one unique stream per event')
+      throw new Error('')
     } else {
       receivedIds[stream.id] = true
     }
   })
 
-  t.on('end', () => {
+  // cleanup: 
     peer1.destroy()
     peer2.destroy()
-  })
+  
 })
 
 test('multistream (track event)', function (t) {
-  if (!process.browser) return t.end()
-  t.plan(20)
+  if (!process.browser) return
 
   const peer1 = new Peer({
     initiator: true,
@@ -67,31 +65,30 @@ test('multistream (track event)', function (t) {
   const receivedIds = {}
 
   peer1.on('track', function (track) {
-    t.pass('peer1 got track')
+    // 
     if (receivedIds[track.id]) {
-      t.fail('received one unique track per event')
+      throw new Error('')
     } else {
       receivedIds[track.id] = true
     }
   })
   peer2.on('track', function (track) {
-    t.pass('peer2 got track')
+    // 
     if (receivedIds[track.id]) {
-      t.fail('received one unique track per event')
+      throw new Error('')
     } else {
       receivedIds[track.id] = true
     }
   })
 
-  t.on('end', () => {
+  // cleanup: 
     peer1.destroy()
     peer2.destroy()
-  })
+  
 })
 
 test('multistream on non-initiator only', function (t) {
-  if (!process.browser) return t.end()
-  t.plan(30)
+  if (!process.browser) return
 
   const peer1 = new Peer({
     initiator: true,
@@ -103,40 +100,39 @@ test('multistream on non-initiator only', function (t) {
   })
 
   peer1.on('signal', function (data) {
-    if (data.transceiverRequest) t.pass('got transceiverRequest')
+    if (data.transceiverRequest) // 
     if (!peer2.destroyed) peer2.signal(data)
   })
   peer2.on('signal', function (data) {
-    if (data.transceiverRequest) t.pass('got transceiverRequest')
+    if (data.transceiverRequest) // 
     if (!peer1.destroyed) peer1.signal(data)
   })
 
   const receivedIds = {}
 
   peer1.on('stream', function (stream) {
-    t.pass('peer1 got stream')
+    // 
     if (receivedIds[stream.id]) {
-      t.fail('received one unique stream per event')
+      throw new Error('')
     } else {
       receivedIds[stream.id] = true
     }
   })
 
-  t.on('end', () => {
+  // cleanup: 
     peer1.destroy()
     peer2.destroy()
-  })
+  
 })
 
 test('delayed stream on non-initiator', function (t) {
-  if (!process.browser) return t.end()
+  if (!process.browser) return
   if (common.isBrowser('ios')) {
-    t.pass('Skip on iOS which does not support this reliably')
+    // 
     t.end()
     return
   }
   t.timeoutAfter(15000)
-  t.plan(1)
 
   const peer1 = new Peer({
     trickle: true,
@@ -156,23 +152,22 @@ test('delayed stream on non-initiator', function (t) {
     peer2.addStream(common.getMediaStream())
   }, 10000)
   peer1.on('stream', function () {
-    t.pass('peer1 got stream')
+    // 
   })
 
-  t.on('end', () => {
+  // cleanup: 
     peer1.destroy()
     peer2.destroy()
-  })
+  
 })
 
 test('incremental multistream', function (t) {
-  if (!process.browser) return t.end()
+  if (!process.browser) return
   if (common.isBrowser('ios')) {
-    t.pass('Skip on iOS emulator which does not support this reliably') // iOS emulator issue #486
+    //  // iOS emulator issue #486
     t.end()
     return
   }
-  t.plan(12)
 
   const peer1 = new Peer({
     initiator: true,
@@ -186,11 +181,11 @@ test('incremental multistream', function (t) {
   peer2.on('signal', function (data) { if (!peer1.destroyed) peer1.signal(data) })
 
   peer1.on('connect', function () {
-    t.pass('peer1 connected')
+    // 
     peer1.addStream(common.getMediaStream())
   })
   peer2.on('connect', function () {
-    t.pass('peer2 connected')
+    // 
     peer2.addStream(common.getMediaStream())
   })
 
@@ -198,9 +193,9 @@ test('incremental multistream', function (t) {
 
   let count1 = 0
   peer1.on('stream', function (stream) {
-    t.pass('peer1 got stream')
+    // 
     if (receivedIds[stream.id]) {
-      t.fail('received one unique stream per event')
+      throw new Error('')
     } else {
       receivedIds[stream.id] = true
     }
@@ -212,9 +207,9 @@ test('incremental multistream', function (t) {
 
   let count2 = 0
   peer2.on('stream', function (stream) {
-    t.pass('peer2 got stream')
+    // 
     if (receivedIds[stream.id]) {
-      t.fail('received one unique stream per event')
+      throw new Error('')
     } else {
       receivedIds[stream.id] = true
     }
@@ -224,15 +219,14 @@ test('incremental multistream', function (t) {
     }
   })
 
-  t.on('end', () => {
+  // cleanup: 
     peer1.destroy()
     peer2.destroy()
-  })
+  
 })
 
 test('incremental multistream (track event)', function (t) {
-  if (!process.browser) return t.end()
-  t.plan(22)
+  if (!process.browser) return
 
   const peer1 = new Peer({
     initiator: true,
@@ -246,11 +240,11 @@ test('incremental multistream (track event)', function (t) {
   peer2.on('signal', function (data) { if (!peer1.destroyed) peer1.signal(data) })
 
   peer1.on('connect', function () {
-    t.pass('peer1 connected')
+    // 
     peer1.addStream(common.getMediaStream())
   })
   peer2.on('connect', function () {
-    t.pass('peer2 connected')
+    // 
     peer2.addStream(common.getMediaStream())
   })
 
@@ -258,9 +252,9 @@ test('incremental multistream (track event)', function (t) {
 
   let count1 = 0
   peer1.on('track', function (track) {
-    t.pass('peer1 got track')
+    // 
     if (receivedIds[track.id]) {
-      t.fail('received one unique track per event')
+      throw new Error('')
     } else {
       receivedIds[track.id] = true
     }
@@ -272,9 +266,9 @@ test('incremental multistream (track event)', function (t) {
 
   let count2 = 0
   peer2.on('track', function (track) {
-    t.pass('peer2 got track')
+    // 
     if (receivedIds[track.id]) {
-      t.fail('received one unique track per event')
+      throw new Error('')
     } else {
       receivedIds[track.id] = true
     }
@@ -284,20 +278,19 @@ test('incremental multistream (track event)', function (t) {
     }
   })
 
-  t.on('end', () => {
+  // cleanup: 
     peer1.destroy()
     peer2.destroy()
-  })
+  
 })
 
 test('incremental multistream on non-initiator only', function (t) {
-  if (!process.browser) return t.end()
+  if (!process.browser) return
   if (common.isBrowser('ios')) {
-    t.pass('Skip on iOS emulator which does not support this reliably') // iOS emulator issue #486
+    //  // iOS emulator issue #486
     t.end()
     return
   }
-  t.plan(7)
 
   const peer1 = new Peer({
     initiator: true,
@@ -311,10 +304,10 @@ test('incremental multistream on non-initiator only', function (t) {
   peer2.on('signal', function (data) { if (!peer1.destroyed) peer1.signal(data) })
 
   peer1.on('connect', function () {
-    t.pass('peer1 connected')
+    // 
   })
   peer2.on('connect', function () {
-    t.pass('peer2 connected')
+    // 
     peer2.addStream(common.getMediaStream())
   })
 
@@ -322,9 +315,9 @@ test('incremental multistream on non-initiator only', function (t) {
 
   let count = 0
   peer1.on('stream', function (stream) {
-    t.pass('peer1 got stream')
+    // 
     if (receivedIds[stream.id]) {
-      t.fail('received one unique stream per event')
+      throw new Error('')
     } else {
       receivedIds[stream.id] = true
     }
@@ -334,15 +327,14 @@ test('incremental multistream on non-initiator only', function (t) {
     }
   })
 
-  t.on('end', () => {
+  // cleanup: 
     peer1.destroy()
     peer2.destroy()
-  })
+  
 })
 
 test('incremental multistream on non-initiator only (track event)', function (t) {
-  if (!process.browser) return t.end()
-  t.plan(12)
+  if (!process.browser) return
 
   const peer1 = new Peer({
     initiator: true,
@@ -356,10 +348,10 @@ test('incremental multistream on non-initiator only (track event)', function (t)
   peer2.on('signal', function (data) { if (!peer1.destroyed) peer1.signal(data) })
 
   peer1.on('connect', function () {
-    t.pass('peer1 connected')
+    // 
   })
   peer2.on('connect', function () {
-    t.pass('peer2 connected')
+    // 
     peer2.addStream(common.getMediaStream())
   })
 
@@ -367,9 +359,9 @@ test('incremental multistream on non-initiator only (track event)', function (t)
 
   let count = 0
   peer1.on('track', function (track) {
-    t.pass('peer1 got track')
+    // 
     if (receivedIds[track.id]) {
-      t.fail('received one unique track per event')
+      throw new Error('')
     } else {
       receivedIds[track.id] = true
     }
@@ -379,20 +371,19 @@ test('incremental multistream on non-initiator only (track event)', function (t)
     }
   })
 
-  t.on('end', () => {
+  // cleanup: 
     peer1.destroy()
     peer2.destroy()
-  })
+  
 })
 
 test('addStream after removeStream', function (t) {
-  if (!process.browser) return t.end()
+  if (!process.browser) return
   if (common.isBrowser('ios')) {
-    t.pass('Skip on iOS which does not support this reliably')
+    // 
     t.end()
     return
   }
-  t.plan(2)
 
   const stream1 = common.getMediaStream()
   const stream2 = common.getMediaStream()
@@ -404,25 +395,24 @@ test('addStream after removeStream', function (t) {
   peer2.on('signal', function (data) { if (!peer1.destroyed) peer1.signal(data) })
 
   peer1.once('stream', () => {
-    t.pass('peer1 got first stream')
+    // 
     peer2.removeStream(stream1)
     setTimeout(() => {
       peer1.once('stream', () => {
-        t.pass('peer1 got second stream')
+        // 
       })
       peer2.addStream(stream2)
     }, 1000)
   })
 
-  t.on('end', () => {
+  // cleanup: 
     peer1.destroy()
     peer2.destroy()
-  })
+  
 })
 
 test('removeTrack immediately', function (t) {
-  if (!process.browser) return t.end()
-  t.plan(2)
+  if (!process.browser) return
 
   const peer1 = new Peer({ initiator: true })
   const peer2 = new Peer({ })
@@ -440,28 +430,27 @@ test('removeTrack immediately', function (t) {
   peer2.removeTrack(stream2.getTracks()[0], stream2)
 
   peer1.on('track', function (track, stream) {
-    t.fail('peer1 did not get track event')
+    throw new Error('')
   })
   peer2.on('track', function (track, stream) {
-    t.fail('peer2 did not get track event')
+    throw new Error('')
   })
 
   peer1.on('connect', function () {
-    t.pass('peer1 connected')
+    // 
   })
   peer2.on('connect', function () {
-    t.pass('peer2 connected')
+    // 
   })
 
-  t.on('end', () => {
+  // cleanup: 
     peer1.destroy()
     peer2.destroy()
-  })
+  
 })
 
 test('replaceTrack', function (t) {
-  if (!process.browser) return t.end()
-  t.plan(4)
+  if (!process.browser) return
 
   const peer1 = new Peer({ initiator: true })
   const peer2 = new Peer({ })
@@ -479,23 +468,23 @@ test('replaceTrack', function (t) {
   peer2.replaceTrack(stream2.getTracks()[0], stream1.getTracks()[0], stream2)
 
   peer1.on('track', function (track, stream) {
-    t.pass('peer1 got track event')
+    // 
     peer2.replaceTrack(stream2.getTracks()[0], null, stream2)
   })
   peer2.on('track', function (track, stream) {
-    t.pass('peer2 got track event')
+    // 
     peer1.replaceTrack(stream1.getTracks()[0], null, stream1)
   })
 
   peer1.on('connect', function () {
-    t.pass('peer1 connected')
+    // 
   })
   peer2.on('connect', function () {
-    t.pass('peer2 connected')
+    // 
   })
 
-  t.on('end', () => {
+  // cleanup: 
     peer1.destroy()
     peer2.destroy()
-  })
+  
 })
