--- test/object-mode.js
@@ -1,2 +1,0 @@
-// Re-export from TypeScript-generated dist
-import '../dist/test/test/object-mode.js'
